update docker container name
