+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/KubernetesDeclarativeAgentTest.java
@@ -29,6 +29,7 @@ import java.util.List;
 import java.util.Map;
 import static org.junit.Assert.*;
 
+import hudson.model.Result;
 import jenkins.plugins.git.GitSampleRepoRule;
 import jenkins.plugins.git.GitStep;
 import org.jenkinsci.plugins.structs.describable.UninstantiatedDescribable;
@@ -138,4 +139,22 @@ public class KubernetesDeclarativeAgentTest extends AbstractKubernetesPipelineTe
         r.assertLogContains("Apache Maven 3.3.9", b);
         r.assertLogNotContains("go version go1.6.3", b);
     }
+
+    @Test
+    public void runWithNonexistentDockerImage() throws Exception {
+        assertNotNull(createJobThenScheduleRun());
+        r.assertBuildStatus(Result.FAILURE, r.waitForCompletion(b));
+        System.out.println("LOG: " + r.getLog(b));
+        r.assertLogContains("ERROR: Unable to pull Docker image", b);
+    }
+
+    @Test
+    public void runWithNonexistentDockerImageLongLabel() throws Exception {
+        assertNotNull(createJobThenScheduleRun());
+        r.assertBuildStatus(Result.FAILURE, r.waitForCompletion(b));
+        System.out.println("LOG: " + r.getLog(b));
+        r.assertLogContains("ERROR: Unable to pull Docker image", b);
+
+    }
+
 }
